Refactor component imports and migrate Leaderboard and Navbar components to a new structure
