+++ src/component/background-beams-with-collision.tsx
